Add ttk/style and tk/font to help with scaling issues. (#5)
Tk notices when a screen is scaled and scales the default fonts
automatically, if I understand correctly. Certainly, on scaled
monitors the fonts are generally readable. Unfortunately, the style
for the ttk::treeview does NOT automatically scale the rowheight
needed. For that you have to manually get the linespace using the Tk
command font and then update the Treeview style using ttk::sstyle.

See the discussion of -rowheight in "STYLING OPTIONS" in the
ttk::treeview documentation.